Split context sections out of Stats Cost sub-tab into a Context sub-tab
